# artifacts/api-server/dist/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetStan Veterinary Dictionary Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload/image:
    post:
      operationId: uploadImage
      tags: [upload]
      summary: Upload an image file
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"

  # --- SEARCH ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UploadResponse:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
      required:
        - url
        - filename

    # --- SEARCH ---
...